+++ packages/react-server/examples/basic/src/routes/test/wasm/page.tsx
@@ -9,7 +9,7 @@ const getHighlither = once(async () => {
     langs: [js],
     // non js extension file is not externalized, so we can transform this via `load` hook
     // https://github.com/vitejs/vite/blob/fcf50c2e881356ea0d725cc563722712a2bf5695/packages/vite/src/node/plugins/resolve.ts#L810-L818
-    loadWasm: import("shiki/onig.wasm" as string),
+    loadWasm: import("shiki/onig.wasm?module" as string),
   });
 });
