+++ packages/react-server/examples/basic/vite.config.ts
@@ -31,10 +31,7 @@ export default defineConfig({
         mdx(),
         testVitePluginVirtual(),
         wasmModulePlugin({
-          mode:
-            process.env.VERCEL || process.env.CF_PAGES
-              ? "asset-import"
-              : "asset-fs",
+          mode: process.env.VERCEL || process.env.CF_PAGES ? "import" : "fs",
         }),
         {
           name: "cusotm-react-server-config",
